+++ .eslintrc.cjs
@@ -5,13 +5,19 @@ module.exports = {
     node: true,
     mocha: true,
   },
-  extends: ["eslint:recommended", "plugin:prettier/recommended"],
+  parser: "@typescript-eslint/parser",
+  extends: [
+    "eslint:recommended",
+    "plugin:@typescript-eslint/eslint-recommended",
+    "plugin:@typescript-eslint/recommended",
+    "plugin:prettier/recommended",
+  ],
   parserOptions: {
     sourceType: "module",
   },
   overrides: [],
   ignorePatterns: ["dist/**/*", "rollup.config.mjs"],
-  plugins: ["prettier", "mocha"],
+  plugins: ["mocha", "@typescript-eslint", "prettier"],
   rules: {
     "mocha/no-exclusive-tests": "error",
   },