+++ src/constants.ts
@@ -1,0 +1,3 @@
+export const PLUGIN_NAME = "keep-css-imports"
+export const FILE_URL_PREFIX = new URL("file://").toString()
+export const KEY_EXT_STRING = ".[keep-css-imports-plugin-ext]"