+++ src/helpers.ts
@@ -1,8 +1,10 @@
-import path from "path"
+import * as path from "path"
+import { EmitFile } from "rollup"
+import { StylesMap } from "./types"
 
 export const escapeRegex = (val) => val.replace(/[/\-\\^$*+?.()|[\]{}]/g, "\\$&")
 
-export const assertDuplicates = (stylesToEmit) => {
+export const assertDuplicates = (stylesToEmit: StylesMap) => {
   Object.values(stylesToEmit).forEach((v, i, all) => {
     if (all.some((av, ai) => !!v.output && v.output === av.output && ai != i)) {
       throw new Error(`Two or more assets have conflicting output path ${v.output}`)
@@ -15,10 +17,15 @@ export const assertLocation = (outDir, assetPath) => {
   }
 }
 
-export const ensureSourceMap = ({ css, map }, includeSourceMap, fileName, onEmit) => {
+export const ensureSourceMap = (
+  { css, map }: { css?: string | Uint8Array; map?: string | Uint8Array },
+  includeSourceMap: boolean | "inline" | undefined,
+  fileName: string,
+  onEmit: EmitFile,
+) => {
   if (map) {
     if (includeSourceMap === "inline") {
-      css += `\n/*# sourceMappingURL=data:application/json;base64,${Buffer.from(map, "utf8").toString("base64")}*/`
+      css += `\n/*# sourceMappingURL=data:application/json;base64,${(map instanceof Uint8Array ? Buffer.from(map) : Buffer.from(map, "utf8")).toString("base64")}*/`
     } else if (includeSourceMap === true) {
       css += `\n/*# sourceMappingURL=${path.basename(fileName)}.map */`
     }
@@ -34,19 +41,19 @@ export const ensureSourceMap = ({ css, map }, includeSourceMap, fileName, onEmit
   return css
 }
 
-export const formatProcessedToCSS = (input, sourceMap) =>
+export const formatProcessedToCSS = (input: string | { css: string; map?: string | object }, sourceMap: boolean) =>
   typeof input === "string"
     ? { css: input, map: "" }
     : typeof input === "object"
-    ? {
-        css: input.css,
-        map: !sourceMap ? "" : typeof input.map === "object" ? JSON.stringify(input.map) : input.map,
-      }
-    : input
+      ? {
+          css: input.css,
+          map: !sourceMap ? "" : typeof input.map === "object" ? JSON.stringify(input.map) : input.map,
+        }
+      : input
 
-export const requireSass = async () => {
+export const requireSass = () => {
   try {
-    return await import("sass")
+    return import("sass")
   } catch (e) {
     throw new Error(
       "You have to install `sass` package! Try running\n\t" +